Move BucketChecksumState to a separate file.
Add more units tests.
