+++ packages/service-core/src/sync/sync-index.ts
@@ -5,3 +5,4 @@ export * from './RequestTracker.js';
 export * from './safeRace.js';
 export * from './sync.js';
 export * from './util.js';
+export * from './BucketChecksumState.js';
